requirements.txt
adal==1.2.2
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
argcomplete==1.10.0
asn1crypto==0.24.0
azure-common==1.1.25
azure-core==1.4.0
azure-functions==1.2.0
azure-multiapi-storage==0.2.4
azure-nspkg==3.0.2
azure-storage==0.33.0
azure-storage-blob==12.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==2.9.2
cycler==0.10.0
et-xmlfile==1.0.1
fabric==2.5.0
humanfriendly==4.18
idna==2.9
invoke==1.3.0
isodate==0.6.0
javaproperties==0.5.1
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.2.0
kiwisolver==1.1.0
knack==0.6.3
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mock==2.0.0
msrest==0.6.13
msrestazure==0.6.1
numpy==1.18.3
oauthlib==3.1.0
openpyxl==2.6.3
pandas==1.0.3
paramiko==2.6.0
pbr==5.4.2
portalocker==1.5.1
psutil==5.6.3
pycparser==2.20
pydocumentdb==2.3.3
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyreadline==2.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
scp==0.13.2
six==1.14.0
soupsieve==1.9.3
sshtunnel==0.1.5
tabulate==0.8.3
urllib3==1.25.9
vsts==0.1.25
vsts-cd-manager==1.0.2
websocket-client==0.56.0
xlrd==1.2.0
xmltodict==0.12.0
